pyproject.toml
[project]
name = "padc"
version = "0.1.0"
description = "Speech-to-text CLI tool with local and cloud transcription"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.0",
    "numpy>=2.3.2",
    "openai>=1.100.2",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.1",
    "sounddevice>=0.5.2",
    "typer>=0.16.1",
]

[project.scripts]
padc = "padc.cli:app"

[project.optional-dependencies]
gpu = [
    "torch>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
